Added separate internal license information link
The hashtags needed to format the link break the link, so I have added an extra comment with the correct link. This new link is only seen by people editing the DDTXT, so it does not matter much, but it is better to make sure people can use the link.