Improved dialogues, added links and summary box
Implemented fix that allows colons to exist in dialog. (Only the first colon is part of DDTXT syntax now.)
Added links through dialog! Use "href" as a character and use the link as the dialogue. (Formatting: [href: link])
Added the summary box, which can be opened through a button at the top of the webpage. This summary box displays the DDT.global.summary attribute that has existed and a new lastModified attribute, which displays the .ddtxt's last modified date as an easy to read string. This summary box also displays the author, which is now displayed three different places.