Vastly improved DDTXT Errors
All errors can now be more dynamically logged. I still have to implement the editor error console so that authors can see the errors, but most of the work is done. In addition, errors are now more standardized instead of being a single line of code using a string made up on the spot.