Added estimate disclaimer to word count
The summary now clarifies that the word count excludes some writing (the summary and character bios) and is an estimation and not perfectly exact. For general purposes though, the current word count should suffice. The current implementation compromises between perfect accuracy and excessive processing for a relatively minute fact that most readers will likely not care about.